Bug 1997101 - Remove redundant type conversions. r=android-reviewers,mcarare
warning: Redundant call of conversion method.

Differential Revision: https://phabricator.services.mozilla.com/D270513

+++ samples/toolbar/src/main/java/org/mozilla/samples/toolbar/ToolbarActivity.kt
@@ -358,10 +358,10 @@ class ToolbarActivity : AppCompatActivity() {
             siteInfoIconSecure = 0xFF20123a.toInt(),
             text = 0xFF0c0c0d.toInt(),
             menu = 0xFF20123a.toInt(),
-            separator = 0x1E15141a.toInt(),
+            separator = 0x1E15141a,
             trackingProtection = 0xFF20123a.toInt(),
             emptyIcon = 0xFF20123a.toInt(),
-            hint = 0x1E15141a.toInt(),
+            hint = 0x1E15141a,
         )
 
         binding.toolbar.display.urlFormatter = { url ->
@@ -446,7 +446,7 @@ class ToolbarActivity : AppCompatActivity() {
             text = 0xFF0c0c0d.toInt(),
             title = 0xFF0c0c0d.toInt(),
             menu = 0xFF20123a.toInt(),
-            separator = 0x1E15141a.toInt(),
+            separator = 0x1E15141a,
             trackingProtection = 0xFF20123a.toInt(),
         )
